client context internal class
