+++ sdk/cosmosdb/cosmos/src/queryExecutionContext/hybridQueryExecutionContext.ts
@@ -3,7 +3,7 @@
 
 import type { AzureLogger } from "@azure/logger";
 import { createClientLogger } from "@azure/logger";
-import type { ClientContext } from "../ClientContext.js";
+import type { ClientContextInternal } from "../ClientContextInternal.js";
 import type { DiagnosticNodeInternal } from "../diagnostics/DiagnosticNodeInternal.js";
 import type {
   FeedOptions,
@@ -48,7 +48,7 @@ export class HybridQueryExecutionContext implements ExecutionContext {
   private isSingleComponent: boolean = false;
 
   constructor(
-    private clientContext: ClientContext,
+    private clientContext: ClientContextInternal,
     private collectionLink: string,
     private options: FeedOptions,
     private partitionedQueryExecutionInfo: PartitionedQueryExecutionInfo,
